+++ src/auth/auth.service.ts
@@ -1,263 +1,97 @@
-import {
-  BadRequestException,
-  Body,
-  Controller,
-  Get,
-  Post,
-  Req,
-  Res,
-} from '@nestjs/common';
-import { Response, Request } from 'express';
-import { randomBytes, randomUUID } from 'node:crypto';
+import { Injectable, UnauthorizedException } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
+import { ConfigService } from '@nestjs/config';
 import * as bcrypt from 'bcrypt';
-
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
+import { IsNull, MoreThan, Repository } from 'typeorm';
 
-import { AzureMsalService } from '@/azure/azure-msal.service';
-import { User } from '@/users/entities/user.entity';
-import { Public } from '@/common/decorators/public.decorator';
-import { AuthUser, RequestUser } from '@/common/decorators/auth-user.decorator';
-import { Role, type AppRole } from '@/common/domain/role.enum';
-import { AuthService } from './auth.controller';
+import type { AppRole } from '../common/domain/role.enum';
+import { Session } from '../auth/entities/session.entity';
 
-interface AzureIdTokenClaims {
-  oid?: string;
-  tid?: string;
-  sub?: string;
-  name?: string;
-  email?: string;
-  preferred_username?: string;
-  emails?: string[];
-}
+type JwtPayload = { sub: number; email: string; role: AppRole };
 
-class RefreshDto {
-  refreshToken?: string;
-}
+const REFRESH_TTL_DAYS = 30;
 
-function sameSite(): 'lax' | 'strict' | 'none' {
-  const v = (process.env.COOKIE_SAMESITE ?? 'lax').toLowerCase();
-  return v === 'none' ? 'none' : v === 'strict' ? 'strict' : 'lax';
-}
-function cookieBase() {
-  return {
-    httpOnly: true as const,
-    secure: process.env.COOKIE_SECURE === 'true',
-    sameSite: sameSite(),
-    domain: process.env.COOKIE_DOMAIN || undefined,
-  };
-}
-function setRefreshCookie(res: Response, token: string) {
-  res.cookie('refresh_token', token, {
-    ...cookieBase(),
-    path: '/auth',
-    maxAge: 30 * 24 * 3600 * 1000,
-  });
-}
-function setSidCookie(res: Response, sid: string) {
-  res.cookie('sid', sid, {
-    ...cookieBase(),
-    path: '/auth',
-    maxAge: 30 * 24 * 3600 * 1000,
-  });
-}
-function setCsrfCookie(res: Response, token: string) {
-  res.cookie('csrf_token', token, {
-    httpOnly: false,
-    secure: process.env.COOKIE_SECURE === 'true',
-    sameSite: sameSite(),
-    domain: process.env.COOKIE_DOMAIN || undefined,
-    path: '/',
-    maxAge: 30 * 24 * 3600 * 1000,
-  });
-}
-function clearAuthCookies(res: Response) {
-  const base = cookieBase();
-  res.cookie('refresh_token', '', { ...base, path: '/auth', maxAge: 0 });
-  res.cookie('sid', '', { ...base, path: '/auth', maxAge: 0 });
-  res.cookie('csrf_token', '', {
-    httpOnly: false,
-    secure: base.secure,
-    sameSite: base.sameSite,
-    domain: base.domain,
-    path: '/',
-    maxAge: 0,
-  });
-}
-function assertCsrfAndOrigin(req: Request) {
-  const csrfCookie = req.cookies?.csrf_token as string | undefined;
-  const csrfHeader = (req.get('x-csrf-token') || req.get('x-xsrf-token')) as
-    | string
-    | undefined;
-  if (!csrfCookie || !csrfHeader || csrfCookie !== csrfHeader) {
-    throw new BadRequestException('Bad CSRF token');
-  }
-  const origin = req.get('origin') || '';
-  const allowed = (process.env.CLIENT_URL || '')
-    .split(',')
-    .map((s) => s.trim())
-    .filter(Boolean);
-  if (origin && allowed.length && !allowed.includes(origin)) {
-    throw new BadRequestException('Bad Origin');
-  }
-}
-
-@Controller('auth')
-export class AuthController {
+@Injectable()
+export class AuthService {
   constructor(
-    private readonly auth: AuthService,
-    private readonly msal: AzureMsalService,
-    @InjectRepository(User) private readonly users: Repository<User>,
+    private readonly jwt: JwtService,
+    private readonly cfg: ConfigService,
+    @InjectRepository(Session) private readonly sessions: Repository<Session>,
   ) {}
 
-  @Public()
-  @Get('azure/login')
-  async azureLogin(@Res() res: Response) {
-    const url = await this.msal.buildAuthUrl();
-    return res.redirect(url);
+  async signAccessToken(user: { id: number; email: string; role: AppRole }) {
+    const payload: JwtPayload = { sub: user.id, email: user.email, role: user.role };
+    return this.jwt.signAsync(payload); // секрет/ttl берутся из JwtModule
   }
 
-  @Public()
-  @Get('azure/callback')
-  async azureCallback(@Req() req: Request, @Res() res: Response) {
-    const currentUrl = new URL(
-      `${req.protocol}://${req.get('host')}${req.originalUrl}`,
-    );
-    const result = await this.msal.exchangeCode(currentUrl);
-
-    const claims = (result.idTokenClaims ?? {}) as AzureIdTokenClaims;
-    const azureId = claims.oid ?? claims.sub;
-    if (!azureId) throw new BadRequestException('Azure ID (oid/sub) not found');
-
-    const email =
-      claims.email ??
-      claims.preferred_username ??
-      (Array.isArray(claims.emails) ? claims.emails[0] : undefined);
-    const displayName = claims.name ?? null;
-    const azureTenantId = claims.tid ?? null;
-
-    // upsert user (по azureId, иначе по email)
-    let user = await this.users.findOneBy({ azureId });
-
-    if (user) {
-      user.displayName = displayName ?? user.displayName;
-      user.azureTenantId = azureTenantId ?? user.azureTenantId;
-      user = await this.users.save(user);
-    } else if (email) {
-      const byEmail = await this.users.findOneBy({ email });
-      if (byEmail) {
-        byEmail.azureId = azureId;
-        byEmail.azureTenantId = azureTenantId ?? byEmail.azureTenantId;
-        byEmail.displayName = displayName ?? byEmail.displayName;
-        user = await this.users.save(byEmail);
-      }
-    }
-
-    if (!user) {
-      user = await this.users.save(
-        this.users.create({
-          email: email,
-          password: await bcrypt.hash(randomUUID(), 12), 
-          role: Role.user,
-          displayName,
-          azureId,
-          azureTenantId,
-        }) as User,
-      );
-    }
-
-    const jwtUser = {
-      id: user!.id,
-      email: user.email,
-      role: user.role as AppRole,
-    };
-
-    const accessToken = await this.auth.signAccessToken(jwtUser);
-    const refreshToken = await this.auth.signRefreshToken(jwtUser);
-    const sessionId = await this.auth.createSession(
-      user.id,
-      refreshToken,
-      req.headers['user-agent']?.toString(),
-      (req.headers['x-forwarded-for'] as string) ||
-        req.socket.remoteAddress ||
-        undefined,
-    );
-
-    setRefreshCookie(res, refreshToken);
-    setSidCookie(res, sessionId);
-    const csrfToken = randomBytes(32).toString('base64url');
-    setCsrfCookie(res, csrfToken);
-
-    return res.json({
-      ok: true,
-      user: {
-        id: user.id,
-        email: user.email,
-        role: user.role,
-        displayName: user.displayName,
-      },
-      access_token: accessToken,
-      expires_in: 15 * 60,
+  async signRefreshToken(user: { id: number; email: string; role: AppRole }) {
+    const payload: JwtPayload = { sub: user.id, email: user.email, role: user.role };
+    const refreshSecret = this.cfg.get<string>('JWT_REFRESH_SECRET')!;
+    return this.jwt.signAsync(payload, {
+      secret: refreshSecret,
+      expiresIn: `${REFRESH_TTL_DAYS}d`,
     });
   }
 
-  @Post('profile')
-  profile(@AuthUser() user: RequestUser) {
-    return user; // { sub, email, role }
+  async verifyRefreshToken(token: string) {
+    const refreshSecret = this.cfg.get<string>('JWT_REFRESH_SECRET')!;
+    return this.jwt.verifyAsync<JwtPayload>(token, { secret: refreshSecret });
   }
 
-  @Public()
-  @Post('refresh')
-  async refresh(
-    @Req() req: Request,
-    @Res() res: Response,
-    @Body() body: RefreshDto,
+  private sessionExpiryDate(): Date {
+    return new Date(Date.now() + REFRESH_TTL_DAYS * 24 * 60 * 60 * 1000);
+  }
+
+  async createSession(
+    userId: number,
+    refreshToken: string,
+    userAgent?: string,
+    ip?: string,
   ) {
-    assertCsrfAndOrigin(req);
-
-    const sid = req.cookies?.sid as string | undefined;
-    const oldRefresh =
-      (req.cookies?.refresh_token as string | undefined) ?? body.refreshToken;
-    if (!sid || !oldRefresh)
-      throw new BadRequestException('No session or refresh token');
-
-    const payload = await this.auth.verifyRefreshToken(oldRefresh);
-    const dbUser = await this.users.findOneOrFail({
-      where: { id: payload.sub },
-      select: { id: true, email: true, role: true },
+    const refreshHash = await bcrypt.hash(refreshToken, 12);
+    const entity = this.sessions.create({
+      userId,
+      refreshHash,
+      userAgent: userAgent ?? null,
+      ip: ip ?? null,
+      expiresAt: this.sessionExpiryDate(),
     });
-
-    const jwtUser = {
-      id: dbUser.id,
-      email: dbUser.email,
-      role: dbUser.role as AppRole,
-    };
-
-    const rotated = await this.auth.rotateSession(
-      sid,
-      oldRefresh,
-      jwtUser,
-      req.headers['user-agent']?.toString(),
-      (req.headers['x-forwarded-for'] as string) ||
-        req.socket.remoteAddress ||
-        undefined,
-    );
-    setRefreshCookie(res, rotated.refreshToken);
-
-    return res.json({
-      ok: true,
-      access_token: rotated.accessToken,
-      expires_in: 15 * 60,
-    });
+    const saved = await this.sessions.save(entity);
+    return saved.id;
   }
 
-  @Public()
-  @Post('logout')
-  async logout(@Req() req: Request, @Res() res: Response) {
-    assertCsrfAndOrigin(req);
-    const sid = req.cookies?.sid as string | undefined;
-    if (sid) await this.auth.revokeSession(sid);
-    clearAuthCookies(res);
-    return res.json({ ok: true });
+  async rotateSession(
+    sessionId: string,
+    oldRefreshToken: string,
+    user: { id: number; email: string; role: AppRole },
+    userAgent?: string,
+    ip?: string,
+  ) {
+    const session = await this.sessions.findOne({
+      where: { id: sessionId, revokedAt: IsNull(), expiresAt: MoreThan(new Date()) },
+    });
+    if (!session) throw new UnauthorizedException('Session not found or expired');
+
+    const match = await bcrypt.compare(oldRefreshToken, session.refreshHash);
+    if (!match) throw new UnauthorizedException('Invalid refresh token');
+
+    const accessToken = await this.signAccessToken(user);
+    const newRefresh = await this.signRefreshToken(user);
+
+    session.refreshHash = await bcrypt.hash(newRefresh, 12);
+    session.userAgent = userAgent ?? null;
+    session.ip = ip ?? null;
+    session.expiresAt = this.sessionExpiryDate();
+    await this.sessions.save(session);
+
+    return { accessToken, refreshToken: newRefresh };
+  }
+
+  async revokeSession(sessionId: string) {
+    await this.sessions.update(
+      { id: sessionId, revokedAt: IsNull() },
+      { revokedAt: new Date() },
+    );
   }
 }
