+++ src/common/interceptors/rotate-on-mutation.interceptor.ts
@@ -8,16 +8,20 @@ import {
 } from '@nestjs/common';
 import type { Request, Response } from 'express';
 import { Observable } from 'rxjs';
-import { mergeMap } from 'rxjs/operators';
-import { AuthService } from '../../auth/auth.service';
-import { PrismaService } from '../../../prisma/prisma.service';
+  import { mergeMap } from 'rxjs/operators';
+import { InjectRepository } from '@nestjs/typeorm';
+import { Repository } from 'typeorm';
+
+import { AuthService } from '@/auth/auth.service';
+import { User } from '@/users/entities/user.entity';
+import type { AppRole } from '@/common/domain/role.enum';
 
 function isMutating(method: string) {
   const m = method.toUpperCase();
   return m === 'POST' || m === 'PUT' || m === 'PATCH' || m === 'DELETE';
 }
 
-function cookieOpts(kind: 'refresh' | 'sid') {
+function cookieOpts() {
   const secure = process.env.COOKIE_SECURE === 'true';
   const sameSite = (process.env.COOKIE_SAMESITE as 'lax' | 'strict' | 'none') ?? 'lax';
   const domain = process.env.COOKIE_DOMAIN || undefined;
@@ -29,7 +33,7 @@ function cookieOpts(kind: 'refresh' | 'sid') {
 export class RotateOnMutationInterceptor implements NestInterceptor {
   constructor(
     private readonly auth: AuthService,
-    private readonly prisma: PrismaService,
+    @InjectRepository(User) private readonly users: Repository<User>,
   ) {}
 
   intercept(context: ExecutionContext, next: CallHandler): Observable<any> {
@@ -37,9 +41,10 @@ export class RotateOnMutationInterceptor implements NestInterceptor {
     const res = context.switchToHttp().getResponse<Response>();
 
     if (!isMutating(req.method)) {
-      return next.handle(); 
+      return next.handle();
     }
 
+    // CSRF + Origin
     const csrfCookie = req.cookies?.csrf_token as string | undefined;
     const csrfHeader = (req.get('x-csrf-token') || req.get('x-xsrf-token')) as string | undefined;
     if (!csrfCookie || !csrfHeader || csrfCookie !== csrfHeader) {
@@ -54,6 +59,7 @@ export class RotateOnMutationInterceptor implements NestInterceptor {
       throw new ForbiddenException('Bad Origin');
     }
 
+    // Session cookies
     const sid = req.cookies?.sid as string | undefined;
     const oldRefresh = req.cookies?.refresh_token as string | undefined;
     if (!sid || !oldRefresh) {
@@ -62,8 +68,9 @@ export class RotateOnMutationInterceptor implements NestInterceptor {
 
     return next.handle().pipe(
       mergeMap(async (data) => {
+        // верификация refresh и подтягивание актуального пользователя
         const payload = await this.auth.verifyRefreshToken(oldRefresh);
-        const user = await this.prisma.user.findUniqueOrThrow({
+        const dbUser = await this.users.findOneOrFail({
           where: { id: payload.sub },
           select: { id: true, email: true, role: true },
         });
@@ -71,12 +78,13 @@ export class RotateOnMutationInterceptor implements NestInterceptor {
         const rotated = await this.auth.rotateSession(
           sid,
           oldRefresh,
-          user,
+          { id: dbUser.id, email: dbUser.email, role: dbUser.role as AppRole },
           req.headers['user-agent']?.toString(),
           (req.headers['x-forwarded-for'] as string) || req.socket.remoteAddress || undefined,
         );
 
-        res.cookie('refresh_token', rotated.refreshToken, cookieOpts('refresh'));
+        res.cookie('refresh_token', rotated.refreshToken, cookieOpts());
+
         if (data && typeof data === 'object' && !Array.isArray(data)) {
           return { ...data, access_token: rotated.accessToken };
         }
