+++ src/users/entities/user.entity.ts
@@ -15,22 +15,42 @@ export class User {
   @PrimaryGeneratedColumn()
   id!: number;
 
-  @Column({ length: 255 }) email!: string;
-  @Column({ length: 255 }) password!: string;
-  @Column({ length: 255, nullable: true }) displayName!: string | null;
+  @Column({ type: 'varchar', length: 255 })
+  email!: string;
+
+  @Column({ type: 'varchar', length: 255 })
+  password!: string;
+
+  @Column({ type: 'varchar', length: 255, nullable: true })
+  displayName!: string | null;
 
   @Column({ type: 'enum', enum: Role, default: Role.user })
   role!: Role;
 
-  @Column({ length: 255, nullable: true }) azureId!: string | null;
-  @Column({ length: 255, nullable: true }) azureTenantId!: string | null;
-  @Column({ length: 500, nullable: true }) avatarUrl!: string | null;
+  @Column({ type: 'varchar', length: 255, nullable: true })
+  azureId!: string | null;
 
-  @OneToMany(() => Session, (s) => s.user) sessions!: Session[];
-  @OneToMany(() => Ticket, (t) => t.author) tickets!: Ticket[];
-  @OneToMany(() => Ticket, (t) => t.assignee) assigned!: Ticket[];
-  @OneToMany(() => Commentary, (c) => c.author) comments!: Comment[];
+  @Column({ type: 'varchar', length: 255, nullable: true })
+  azureTenantId!: string | null;
 
-  @CreateDateColumn() createdAt!: Date;
-  @UpdateDateColumn() updatedAt!: Date;
+  @Column({ type: 'varchar', length: 500, nullable: true })
+  avatarUrl!: string | null;
+
+  @OneToMany(() => Session, (s) => s.user)
+  sessions!: Session[];
+
+  @OneToMany(() => Ticket, (t) => t.author)
+  tickets!: Ticket[];
+
+  @OneToMany(() => Ticket, (t) => t.assignee)
+  assigned!: Ticket[];
+
+  @OneToMany(() => Commentary, (c) => c.author)
+  comments!: Commentary[];
+
+  @CreateDateColumn({ type: 'datetime' })
+  createdAt!: Date;
+
+  @UpdateDateColumn({ type: 'datetime' })
+  updatedAt!: Date;
 }
